Add table list and title editor
This is a crude mechanism to edit the metadata for the table.  Right now this only includes the title.  This also adds a list of current tables in the database with a button to choose one.
Other: adds a build task for TSC watch mode for the app.  Removes the index file from the src folder.
Remaining: can get rid of gulp at this point.  Clean up the code and start thinking about features.