Add global shortcut and delete ability
This adds a global shortcut that will show the app and open the search bar.
Adds the ability to delete a Table with the button.
Changed to using Bootstrap 3 instead of 4 (issue with labels).
Wrote a wrapper around the docs creation so that they are of the correct TypeScript type (and have the correct methods available).
This sets the stage for importing more data and then enabling searching.